fix: dependencies, discovery queue
