+++ package.json
@@ -14,20 +14,20 @@
 		"@sveltejs/adapter-auto": "next",
 		"@sveltejs/kit": "next",
 		"autoprefixer": "^10.4.13",
-		"eslint": "^8.28.0",
+		"eslint": "^8.30.0",
 		"eslint-config-prettier": "^8.5.0",
 		"eslint-plugin-svelte3": "^4.0.0",
-		"postcss": "^8.4.19",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.8.1",
-		"svelte": "^3.53.1",
-		"svelte-preprocess": "^4.10.7",
+		"postcss": "^8.4.20",
+		"prettier": "^2.8.1",
+		"prettier-plugin-svelte": "^2.9.0",
+		"svelte": "^3.55.0",
+		"svelte-preprocess": "^5.0.0",
 		"tailwindcss": "^3.2.4",
-		"vite": "^3.2.4"
+		"vite": "^4.0.2"
 	},
 	"type": "module",
 	"dependencies": {
-		"mongodb": "^4.12.1",
-		"mongoose": "^6.7.4"
+		"mongodb": "^4.13.0",
+		"mongoose": "^6.8.1"
 	}
 }
